Create GroupCounselingPrompt
This prompt can be used to create group lessons/sessions for counseling.  It can easily be modified for other group sessions.  I have ran it more than 10 times and it is working smoothly.  If it doesn't show you a handout.  Just write "Show handout" as a second prompt but I don't think that is an issue with the updated prompt.